requirements.lock.txt
aiofiles==25.1.0
annotated-types==0.7.0
anyio==4.11.0
certifi==2026.2.25
cffi==2.0.0
chardet==7.4.3
click==8.3.0
colorama==0.4.6
cryptography==46.0.2
distro==1.9.0
ecdsa==0.19.1
et_xmlfile==2.0.0
fastapi==0.119.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
MarkupSafe==3.0.3
numpy==2.3.3
openai==2.30.0
openpyxl==3.1.5
pandas==2.3.3
passlib==1.7.4
pillow==12.1.1
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.0
pydantic_core==2.41.1
PyMuPDF==1.27.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
qrcode==8.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.3
uvicorn==0.37.0
volcengine-python-sdk==5.0.21
watchfiles==1.1.1
websockets==15.0.1
xlrd==2.0.2
